Extract AbstractRange interface
